Revision to the dataset, R markdown and checklist.
In the dataset: I revised a few numbers as there was a new poster for the two phase III trials. There are no exact SDs, and information about dropouts. Thus, the risk of bias can be better evalauted and now have low risk of bias. Need to double-check the data with Oxford RAs.

In the Rmarkdwon: After follow-up with Virginia, I did not use HK corrections (<5 studies), but this can be changed back if needed. I also re-run the markdown using the updated dataset and updated the SoE tables. Potential refinemenets are needed.

In the checklist: I reported that HK corrections were not done, as <5 studies. I also reported that sensitivity analyses cannot be done as not studies to be excluded (e.g., no studies used imputed values, and no studies had not a low risk of bias).