Added an exploratorative dose-response meta-analysis
A similar method to standardize the doses for ulotaront and rlamitaront with that of hte animal studies was used. The difference is that instead of log, i used Michaelis Menten taken into cosndieration both the Emax and E50. There are generally not enough studies for this (i.e., 3) and not good data for ralmitaront (while for ulotaront I could use Gallupi et al to estimate the concentrations for the daily doses). Thus, this may not be used.